bpf

Module logging

Structs

bpf_map_log

Log for a eBPF map access via a bpf() syscall

generic_msg_log

Generic log with a message field

inode_access_log

Log for various events that access inodes (file_open, inode_unlink)

kernel_load_data_log

Log for a kernel_load_data() LSM hook event

kernel_module_request_log

Log for a kernel_module_request() LSM hook event

kernel_read_file_log

Log for a kernel_read_file() LSM hook event

log_hdr

Header attached to every log message.

ptrace_access_check_log

Log for a ptrace_access_check() LSM hook event

sb_umount_log

Log a sb_umount() syscall

task_kill_log

Log a task_kill() LSM hook event

Enums

EventType

The link between a program’s log structure and the logging system.

LogLevel

Standard log levels indicating the severity of the message.

LogReason

Standard reasons as to why a log is being output.

Constants

BASE_POLICY_ID

BPF_MAP_NAME_LEN

COMM_LEN

MAX_STR_LEN

NO_POL_ID

PATH_MAX

Statics

LOG_FILTER

LOG_LEVEL

Static global to use within ringbuffer callbacks

Functions

char_array_to_str

Utility function which attempts to decode an array of signed bytes (char[] in C) into a UTF-8 string.

log

Outputs a [tracing::Event] according to the log level.

log_header

Attempts to transmute raw bytes into a log_hdr

setup_logger

Builds and reads from an eBPF ringbuffer.